Update to main change and misloot
If you loot an item you cannot use you will be charged based on the avg BRV of one week of raiding. Historically this was x55 BRV (Velious/Luclin) but it is lower now. On raids, auto-loot no drop should be off and you should confirm your loot before accepting

MAIN CHANGES
We understand people may want to change classes to keep the game enjoyable. We are willing to accommodate a wide host of class changes as long as the officers feel you are not abusing the system. Just ask; we are always willing to have that conversation. Very rarely, if at all, will an unapproved main change penalty be given after a conversation with the officer team.

Unapproved:
80% of DKP spent on loot from the last 90ish days or 150x BRV whichever is higher, no exceptions. Officer's decision if a new app is required.

Approved:
Charge of 50% of DKP spent on loot from the last 90ish days or 75x BRV, whichever is lower. Full membership retained.

Officer Requested:
No DKP Penalty. Full membership retained.

These are guidelines. If a unique situation occurs it will be weighed by the officer-core and discussed with the member on an individual basis.

IF YOU DO NOT HAVE ENOUGH DKP TO COVER THE BRV CHARGE, YOU WILL GO NEGATIVE.